Move : move SelectInputContainer & rename SelectAuto #19
